+++ NEWS.md
@@ -1,5 +1,8 @@
 # Changes in mirt 1.46
 
+- Exposed `marginal_moments()` function to compute marginal item/bundle moment 
+  information for the scoring functions
+  
 - Added `itemtype = 'monospline'` for the monotonic spline model 
 
 - `RCI()` gains logical argument `expected.scores` to convert factor scores into metric